+++ dev/TestWrapper.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Component from 'src/Component.jsx';
+import Canvas from 'src/main.jsx';
 
 function drawBackground({ctx, delta, time}) {
   const { width, height } = ctx.canvas;
@@ -53,10 +53,10 @@ export default class TestWrapper extends React.Component {
     return(
       <div>
         <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1, user-scalable=no" />
-        <Component draw={drawBackground} top={100} left={200} width={width} height={height} realtime={true}>
-          <Component draw={drawLeft} top={250} left={100}/>
-          <Component draw={drawRight} top={-100} left={-100}/>
-        </Component>
+        <Canvas draw={drawBackground} top={100} left={200} width={width} height={height} realtime={true}>
+          <Canvas draw={drawLeft} top={250} left={100}/>
+          <Canvas draw={drawRight} top={-100} left={-100}/>
+        </Canvas>
       </div>
     );
   }